Break up file_uploader
